refactor(stack_move): use layout.encoding address wrappers
