+++ python/bloqade/lanes/dialects/stack_move.py
@@ -4,7 +4,7 @@ from bloqade.decoders.dialects.annotate.types import DetectorType, ObservableTyp
 from kirin import ir, lowering, types
 from kirin.decl import info, statement
 
-from bloqade.lanes.bytecode import LaneAddress, LocationAddress, ZoneAddress
+from bloqade.lanes.layout.encoding import LaneAddress, LocationAddress, ZoneAddress
 from bloqade.lanes.types import ArrayType, MeasurementFutureType
 
 dialect = ir.Dialect(name="lanes.stack_move")
